2080_LLM/Psy_Agent_spanBERT_Evidence_Binding/criteria_binder/pyproject.toml
# File: pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "criteria-binder"
version = "0.1.0"
description = "SpanBERT-based evidence binding model for criteria matching"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "torch>=2.0.0",
    "transformers>=4.30.0",
    "datasets>=2.14.0",
    "numpy>=1.24.0",
    "PyYAML>=6.0",
    "tqdm>=4.65.0",
    "scikit-learn>=1.3.0",
    "accelerate>=0.20.0",
    "tokenizers>=0.13.0",
    "hydra-core>=1.3.2",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "black>=23.0.0",
    "isort>=5.12.0",
    "mypy>=1.5.0",
    "types-PyYAML",
]

[tool.setuptools.packages.find]
where = ["src"]

[tool.hydra]
version = "1.3"

[tool.black]
line-length = 88
target-version = ["py310"]

[tool.isort]
profile = "black"
line_length = 88

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true